+++ pyprom/dataexport.py
@@ -2,16 +2,24 @@ import logging
 import os
 from pykml.factory import KML_ElementMaker as KML
 from lxml import etree
+from lib.locations import SpotElevationContainer
 
 
 class KMLfileWriter(object):
     def __init__(self, kmlList, outputFile):
+        """
+        :param kmlList: a list of SpotElevation type objects, or a
+         SpotElevationContainer object
+        :param outputFile: /path/to/your/file.kml
+        """
         self.logger = logging.getLogger('pyProm.{}'.format(__name__))
         self.logger.info("KML Output: {}, {} points".format(outputFile,
                                                             len(kmlList)))
         self.kml = kmlList
         self.outputFile = os.path.expanduser(outputFile)
         self.kmlPoints = list()
+        if isinstance(kmlList, SpotElevationContainer):
+            kmlList = kmlList.points
         for spotElevation in kmlList:
             self.kmlPoints.append(self.generateKMLPlacemark(spotElevation))
         self.kmlFolder = KML.folder(*self.kmlPoints)